+++ src/routes/user.ts
@@ -1,17 +1,36 @@
-import { Router, Request, Response, NextFunction } from 'express';
+import { Router,  Response, NextFunction, Request } from 'express';
 import UserController from '../controllers/user';
 import { globalErrorHandler } from '../utils/errorHandler';
+import { authenticator } from '../middleware/authenticator';
 
 
 const router = Router();
-const userController = new UserController(); // Create an instance of the controller
+const userController = new UserController();
 
-router.get('/', (req: Request, res: Response) => {
-  res.json({ message: 'User Route' });
+// GET request with authenticator middleware
+router.get('/', authenticator, async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    await userController.get(req, res); // Use the controller method
+  } catch (error) {
+    next(error); // Pass the error to the global error handler
+  }
 });
 
+// POST request to create a user
 router.post('/', (req: Request, res: Response, next: NextFunction) => {
-  userController.create(req, res); // Use the instance method here
+  userController.create(req, res).catch(next); // Catch errors and pass to next handler
+}, globalErrorHandler);
+
+router.post('/login', (req: Request, res: Response, next: NextFunction) => {
+  userController.login(req, res).catch(next);
+}, globalErrorHandler)
+
+router.get('/farmers', authenticator, async (req: Request, res: Response, next: NextFunction) => {
+  try {
+    await userController.findFarmers(req, res); // Use the controller method
+  } catch (error) {
+    next(error); // Pass the error to the global error handler
+  }
 });
 
 export default router;