+++ src/utils/errorHandler.ts
@@ -1,4 +1,5 @@
 import { Request, Response, NextFunction } from "express";
+import Logger from "./errorlogger";
 
 // Custom error class for throwing errors
 class AppError extends Error {
@@ -9,9 +10,14 @@ class AppError extends Error {
     }
 }
 
+
+const logger = Logger.getInstance();
+
 export const globalErrorHandler = (err: any, res: Response) => {
-    console.log(err);
     const statusCode = err.statusCode || 500;
+    if (statusCode === 500) {
+        logger.logError(`${!err.statusCode ? 'Spotted Without Status Code ' : ' '}${err}`);
+    }
     const message = err.message || 'Internal Server Error';
     res.status(statusCode).json({ error: message });
 };