+++ changelog.md
@@ -1,7 +1,13 @@
-# 9 Jan 2025
+### 9 Jan 2025
+
+**Updated**
+- [`09 Jan 2025 | Evening`] :  User's Location field has been changed from Object with only `latitude` & `longitude` to Object with type and coordinates.
 
 **Added**
-- User Routes and minimum system setup
+- Stock Routes, User Routes and nearby stocks route
+- dynamic routes manager for APIs and Logging Middleware for error logging
+- functionalities to add and remove stock from user
+- renderer.ts for error logs to markdown file conversion
 - Prefer /apis/api.md for API documentation
 - Specific `Types`, `interface` and `enum` for each entity and functions
 - readme.md file for backend setup
@@ -12,4 +18,5 @@
 **Removed**
 - nodemon support for development
 
-- [github | Jeevan Joshi](https://github.com/jeevanjoshi4434)
+**Contributors**
+[github | Jeevan Joshi](https://github.com/jeevanjoshi4434)